src_django/pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
description = "This project has developed for an interview task"
authors = ["Ali Davood <alidavood.it@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
poetry = "1.7.1"
djangorestframework = "3.14.0"
drf-yasg = "1.21.7"
ipython = "8.21.0"
django-environ = "0.11.2"
psycopg2-binary = "2.9.9"
django-redis = "5.4.0"
factory-boy = "3.2.1"
gunicorn = "21.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
